requirements.txt
# SuperClawBot - Robot Controller Dependencies
# Python 3.8+ required

# GUI Framework
PySide6>=6.5.0

# Numerical Computing
numpy>=1.21.0

# Audio Processing
sounddevice>=0.4.6

# Signal Processing
scipy>=1.7.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Machine Learning
tflite-runtime>=2.10.0

# Serial/Bluetooth Communication
pyserial>=3.5

# Bluetooth Support (Optional - for device discovery)
# Note: PyBluez installation varies by platform:
# - Linux: pip install pybluez
# - Windows: pip install pybluez-win10 (requires Microsoft C++ Build Tools)
# - macOS: pip install pybluez (requires Xcode Command Line Tools)
# If PyBluez fails to install, the app will still work with:
# - Virtual Bluetooth mode (for testing)
# - Direct serial port connections
# - Pre-paired devices via system Bluetooth settings
pybluez>=0.23; platform_system == "Linux"

